Added a logOnSuccess setting
On production servers the rollbar library prints out

    [Rollbar] Successful api response

every time a message is successfully sent. Since overwhelmingly this is expected to succeed it seems redundant to have it in place. I've added the logOnSuccess setting (and defaulted it to false) so this is suppressed by default (and can be enabled if desired). This doesn't affect any of the existing error logging, but reduces the number of extra log lines that simply say "yes, I did what you wanted and nothing went wrong".