Check LRU cache for file context
The `shouldReadFrameFile` function uses the LRU cache to determine whether
the file has been cached or not, but the cached data is only read from the
closure variable `tempFileCache`.

The easiest way to see the effects of this is to report the same error twice
in a row. The first report will contain the stack annotations, but the second
report will not, since the data is contained in the LRU cache but not in
`tempFileCache`. Files contained in the LRU cache will not be annotated in any
stack trace regardless of the error.