Log messages with debug package; remove verbose option
Instead of plain console.log and console.error calls, this patch
adds a logger module which provides equivalent log and error
functions.

These are instances of the debug package, and automatically add a
coloured "Rollbar" prefix, and some timing information.

These messages are namespaced as Rollbar:log and Rollbar:error, and
can be enabled/disabled by the user with the DEBUG environment
variable (which can include `Rollbar:*` for all). The `verbose` option
is therefore surplus and has been removed.