Use rewire to fix uninitialize test
The test was being run in the same process as the previous tests
which meant that rollbar.init() had been called at least once.
Since this test is meant to check the case where init() is not
called at all, we need to use rewire to reset the state of the
notifier module.

In the future, notifiers will store their state in instance vars
and not globally which will make this easier to test.

@jondeandres